RF: moved debian specific retracing tests under test_debian
